TASK: Introduce `TestingViewForFusionRuntime` for testing instead of using a hacky anonymous class
